Safer handling of getopt's optarg variable.
Specifically, always strdup() it -- do not assume that keeping a
long-lived pointer to some other library's global variable is safe. On
some platforms, some of the time, it's not: see
  https://stackoverflow.com/questions/15523328/assigning-to-the-pointer-which-is-a-member-of-the-struct-causes-segmentation-fau

This meant a little rearranging of other code dealing with the format
string to ensure it is always on the heap, so can always be free()d at
exit.